FIxed color size issue for basic animation. Added facilities to handle animations and a single, static color animation. Pointers and vectors and brain bleeds oh my.

+++ lib/NeoPico/src/NeoPico.cpp
@@ -44,7 +44,7 @@ void NeoPico::Clear() {
   }
 }
 
-void NeoPico::SetFrame(uint8_t newFrame[100]) {
+void NeoPico::SetFrame(uint32_t newFrame[100]) {
   for (int i = 0; i < 100; i++) {
     this->frame[i] = newFrame[i];
   }
